refactor: moved Signer from @taquito/taquito to @taquito/core to avoid circular dependencies
