+++ packages/taquito-core/src/signer-interfaces.ts
