+++ packages/taquito-core/src/taquito-core.ts
@@ -4,3 +4,4 @@
  */
 
 export * from './errors';
+export * from './signer-interfaces';
