+++ packages/taquito-signer/package.json
@@ -74,7 +74,6 @@
     "@stablelib/pbkdf2": "^1.0.1",
     "@stablelib/sha512": "^1.0.1",
     "@taquito/core": "^23.0.0-RC.0",
-    "@taquito/taquito": "^23.0.0-RC.0",
     "@taquito/utils": "^23.0.0-RC.0",
     "@types/bn.js": "^5.1.5",
     "bip39": "3.1.0",
