+++ packages/taquito-signer/src/taquito-signer.ts
@@ -18,9 +18,14 @@ import pbkdf2 from 'pbkdf2';
 import * as Bip39 from 'bip39';
 import { Curves, generateSecretKey } from './helpers';
 import { InvalidMnemonicError, InvalidPassphraseError } from './errors';
-import { InvalidKeyError, ProhibitedActionError } from '@taquito/core';
+import {
+  InvalidKeyError,
+  ProhibitedActionError,
+  SignResult,
+  RawSignResult,
+  Signer,
+} from '@taquito/core';
 import { SigningKey, isPOP, PublicKey } from './key-interface';
-import { SignResult, RawSignResult, Signer } from '@taquito/taquito';
 import { BLSKey, BLSPublicKey } from './bls-key';
 
 export * from './import-key';
