+++ packages/taquito/src/taquito.ts
@@ -18,7 +18,7 @@ import { RpcPacker } from './packer/rpc-packer';
 import { TzReadProvider } from './read-provider/interface';
 import { RpcReadAdapter } from './read-provider/rpc-read-adapter';
 import { PreparationProvider } from './prepare/interface';
-import { Signer } from './signer/interface';
+import { Signer } from '@taquito/core';
 import { NoopSigner } from './signer/noop';
 import { SubscribeProvider } from './subscribe/interface';
 import { PollingSubscribeProvider } from './subscribe/polling-subcribe-provider';
@@ -45,7 +45,7 @@ export { CompositeForger } from './forger/composite-forger';
 export { RpcForger } from './forger/rpc-forger';
 export * from './operations';
 export { OperationBatch } from './batch/rpc-batch-provider';
-export * from './signer/interface';
+export { Signer } from '@taquito/core';
 export * from './subscribe/interface';
 export { SubscribeProvider } from './subscribe/interface';
 export { PollingSubscribeProvider } from './subscribe/polling-subcribe-provider';
